+++ package.json
@@ -6,15 +6,18 @@
     "@testing-library/jest-dom": "^5.11.10",
     "@testing-library/react": "^11.2.6",
     "@testing-library/user-event": "^12.8.3",
+    "bootstrap": "^5.0.2",
     "firebase": "^8.6.5",
     "react": "^17.0.2",
+    "react-bootstrap": "^1.6.1",
+    "react-calendar": "^3.4.0",
     "react-dom": "^17.0.2",
     "react-router": "^5.2.0",
     "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.3",
     "semantic-ui-css": "^2.4.1",
     "semantic-ui-react": "^2.0.3",
-    "tailwindcss": "^2.1.1",
+    "tailwindcss": "^2.2.4",
     "uuid": "^8.3.2",
     "web-vitals": "^1.1.1"
   },
