+++ src/Recipes.jsx
@@ -1,43 +1,57 @@
 import React, { useState, useEffect, useContext } from 'react'
+import RecipeCard from './components/RecipeCard'
+import RecipeForm from './RecipeForm'
 import { Container } from 'semantic-ui-react'
 import './Recipes.css'
-
-import { useRecipeContext, useRecipeUpdateContext } from './RecipeContext'
-
-
+import * as RecipeContext from './RecipeContext'
 
 export default function Recipes() {
-  const recipes = useRecipeContext()
-  const updateRecipes = useRecipeUpdateContext()
+  const initState = { title: '', description: '', ingredients: [] }
+  const [recipe, setRecipe] = useState(initState)
 
-  useEffect(() => {
-    //another test to make sure its working
-    if (recipes) {
-      setMyRecipes(recipes)
-    }
+  const recipes = RecipeContext.useRecipeContext()
+  const editRecipe = RecipeContext.useRecipeUpdateContext()
+  const addRecipe = RecipeContext.useRecipeAddContext()
+  const deleteRecipe = RecipeContext.useRecipeDeleteContext()
 
-  }, [recipes])
+  const handleAddRecipe = (recipe) => {
+    addRecipe(recipe)
+    setRecipe(initState)
+  }
 
-  const [myRecipes, setMyRecipes] = useState(null)
-  const [recipe, setRecipe] = useState(null)
+  const handleEditRecipe = (recipe) => {
+    editRecipe(recipe)
+    setRecipe(initState)
+  }
+
+  const handleDeleteRecipe = (recipe) => {
+    deleteRecipe(recipe)
+    setRecipe(initState)
+  }
+
+  const clearRecipe = () => {
+    setRecipe(initState)
+  }
+
+  const handleChange = e => {
+    const { name, value } = e.target
+    setRecipe(prev => ({ ...prev, [name]: value }))
+  }
+
+
   return (
     <Container className="recipePageBody">
       <div>
-        <h1>Hello</h1>
-        <ul>
-          {myRecipes?.map((recipe, i) => {
-            console.log(recipe)
-
-            return (<li key={i} onClick={() => { setRecipe(recipe) }}>{recipe.name}</li>)
-          })
-          }
-        </ul>
-
-        <span>
-          {recipe?.ingredients.map((ingred, i) => {
-            return ingred
-          })}
-        </span>
+        {
+          recipes.length === 0 ?
+            null :
+            recipes.map((recipe, i) => <h1 key={i} onClick={() => setRecipe(recipe)}><RecipeCard recipe={recipe} /> </h1>)
+        }
+        <RecipeForm handleChange={handleChange} recipe={recipe} setRecipe={setRecipe} />
+        <br />
+        <button onClick={recipe?.id ? () => handleEditRecipe(recipe) : () => handleAddRecipe(recipe)}>{recipe?.id ? 'Update Recipe' : 'Add Recipe'}</button>
+        <button onClick={clearRecipe}>Clear Recipe</button>
+        <button onClick={() => handleDeleteRecipe(recipe)} disabled={recipe?.id ? false : true}>Delete Recipe</button>
       </div>
 
     </Container>
